move static method and test
